# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  commonerrors.ErrorMap:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  commonerrors.GenericResponseError:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  newsletterDTO.CreateNewsletterRequest:
    properties:
      email_text:
        type: string
      limit:
        type: integer
    required:
    - email_text
    type: object
  newsletterDTO.GenericResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Rasta
  title: Rasta API
  version: "1.0"
paths:
...
  /newsletter/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a subscriber from the newsletter system using the provided
        email address.
      parameters:
      - description: User email
        in: body
        name: email
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted subscriber
          schema:
            $ref: '#/definitions/newsletterDTO.GenericResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.GenericResponseError'
      summary: Delete Subscriber
      tags:
      - Newsletter
  /newsletter/send:
    post:
      consumes:
      - application/json
      description: Sends the newsletter email to all active subscribers.
      parameters:
      - description: Newsletter content and limit
        in: body
        name: newsletter
        required: true
        schema:
          $ref: '#/definitions/newsletterDTO.CreateNewsletterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully sent newsletter to all active participants
          schema:
            $ref: '#/definitions/newsletterDTO.GenericResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.GenericResponseError'
      summary: Send Newsletter to Active Subscribers
      tags:
      - Newsletter
  /newsletter/subscribe:
    post:
      consumes:
      - application/json
      description: Subscribes the user to the newsletter with the provided email address.
      parameters:
      - description: User email
        in: body
        name: email
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Successfully subscribed to newsletter
          schema:
            $ref: '#/definitions/newsletterDTO.GenericResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "406":
          description: Email already subscribed
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Subscribe to Newsletter
      tags:
      - Newsletter
  /newsletter/subscribers:
    get:
      description: Retrieves a list of all active newsletter subscribers.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched subscribers
          schema:
            $ref: '#/definitions/newsletterDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Get Active Subscribers
      tags:
      - Newsletter
  /newsletter/subscribers/count:
    get:
      description: Retrieves the count of active newsletter subscribers.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched subscribers count
          schema:
            $ref: '#/definitions/newsletterDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Get Active Subscribers Count
      tags:
      - Newsletter
  /newsletter/unsubscribe:
    post:
      consumes:
      - application/json
      description: Unsubscribes the user from the newsletter using the provided email
        address.
      parameters:
      - description: User email
        in: body
        name: email
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Successfully unsubscribed from newsletter
          schema:
            $ref: '#/definitions/newsletterDTO.GenericResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "406":
          description: Email not subscribed
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Unsubscribe from Newsletter
      tags:
      - Newsletter
  /newsletter/unsubscribed/count:
    get:
      description: Retrieves the count of unsubscribed users from the newsletter.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully fetched unsubscribed count
          schema:
            $ref: '#/definitions/newsletterDTO.GenericResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/commonerrors.ErrorMap'
      summary: Get Unsubscribed Count
      tags:
      - Newsletter
...
swagger: "2.0"